+++ src/consts/mod.rs
@@ -1,0 +1,2 @@
+pub mod graphics;
+pub mod keyboard;
